+++ src/Renderer.h
@@ -24,7 +24,7 @@ private:
     Vector3f traceRay(const Ray &ray, float tmin, int bounces,
                       Hit &hit, float refIndex) const;
 
-    Vector3f estimatePixel(const Ray &ray, float tmin, int length, int iters);
+    Vector3f estimatePixel(const Ray &ray, float tmin, float length, int iters);
 
     std::vector<Ray> choosePath(const Ray &ray, Object3D *light, float tmin, float length) const;
